Fortran integer round-off bug (#783)
A bug in subroutine unsinterp_integer_apply is due to int(5.9)=5 while nint(5.9)=6. Hence we change Fortran function int() to nint() for 
M src/mpasjedi/Fields/mpas_fields_mod.F90
M src/mpasjedi/Utilities/mpasjedi_unstructured_interp_mod.F90

3denvar-60-iter_OIE120km run for the first DA step shows identical jedi.log file with benchmark, indicating the modified code is not executed in the run.

+++ src/mpasjedi/Fields/mpas_fields_mod.F90
@@ -1026,7 +1026,7 @@ subroutine deserialize_fields(self, vsize, vect_inc, index)
             if (poolItr % dataType == MPAS_POOL_INTEGER) then
                call mpas_pool_get_array(self%subFields, trim(poolItr % memberName), i1d_ptr_a)
                do hh = 1, nhoriz
-                  i1d_ptr_a(hh) = int ( vect_inc(index + 1) )
+                  i1d_ptr_a(hh) = nint ( vect_inc(index + 1) )
                   index = index + 1
                enddo
             else if (poolItr % dataType == MPAS_POOL_REAL) then
@@ -1043,7 +1043,7 @@ subroutine deserialize_fields(self, vsize, vect_inc, index)
                call mpas_pool_get_array(self%subFields, trim(poolItr % memberName), i2d_ptr_a)
                do vv = 1, nvert
                   do hh = 1, nhoriz
-                     i2d_ptr_a(vv, hh) = int ( vect_inc(index + 1) )
+                     i2d_ptr_a(vv, hh) = nint ( vect_inc(index + 1) )
                      index = index + 1
                   enddo
                enddo
